!!! FEATURE: Persist more than the latest scent
Store the scents in the database to be able to detect more than one latest scent.
This should help in scenarios when older pods of a release are starting by the autoscaler while a new version is already beeing deployed.

In addition the `./flow scentmark:cleanup --keep 10` command is added to cleanup the table of scents.